improving app implementation import syntax
