+++ stream_alert/apps/_apps/duo.py
@@ -22,9 +22,7 @@ import urllib
 
 import requests
 
-from stream_alert.apps import LOGGER
-from stream_alert.apps._apps import StreamAlertApp
-from stream_alert.apps._apps.app_base import AppIntegration
+from . import AppIntegration, LOGGER, StreamAlertApp
 
 
 class DuoApp(AppIntegration):
