+++ stream_alert/apps/_apps/onelogin.py
@@ -15,9 +15,7 @@ limitations under the License.
 """
 import re
 
-from stream_alert.apps import LOGGER
-from stream_alert.apps._apps import StreamAlertApp
-from stream_alert.apps._apps.app_base import AppIntegration
+from . import AppIntegration, LOGGER, StreamAlertApp
 
 
 @StreamAlertApp
