+++ stream_alert/apps/_apps/slack.py
@@ -16,9 +16,7 @@ limitations under the License.
 import re
 import time
 
-from stream_alert.apps import LOGGER
-from stream_alert.apps._apps import StreamAlertApp
-from stream_alert.apps._apps.app_base import AppIntegration
+from . import AppIntegration, LOGGER, StreamAlertApp
 
 
 class SlackApp(AppIntegration):
