+++ stream_alert/apps/app_base.py
